refactoring Enums into modules
