+++ src/main/java/demo/codeexample/logger/application/LoggerAdapter.java
@@ -1,10 +1,9 @@
 package demo.codeexample.logger.application;
 
-import demo.codeexample.enums.LoggerAction;
+import demo.codeexample.logger.domain.LoggerAction;
 import demo.codeexample.logger.LoggerPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import org.springframework.web.context.annotation.RequestScope;
 
 @Component
 @RequiredArgsConstructor
