+++ src/main/java/demo/codeexample/logger/domain/LoggerEntity.java
@@ -1,7 +1,5 @@
 package demo.codeexample.logger.domain;
 
-import demo.codeexample.enums.LoggerAction;
-
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
