+++ src/main/java/demo/codeexample/project/domain/Category.java
@@ -1,4 +1,4 @@
-package demo.codeexample.enums;
+package demo.codeexample.project.domain;
 
 public enum Category {
     FILM,
