+++ src/main/java/demo/codeexample/project/domain/Genre.java
@@ -1,4 +1,4 @@
-package demo.codeexample.enums;
+package demo.codeexample.project.domain;
 
 public enum Genre {
     ACTION,
