+++ src/main/java/demo/codeexample/user/domain/Role.java
@@ -1,4 +1,4 @@
-package demo.codeexample.enums;
+package demo.codeexample.user.domain;
 
 public enum Role {
     PRODUCER,
