+++ src/main/java/demo/codeexample/user/domain/UserRepository.java
@@ -1,6 +1,5 @@
 package demo.codeexample.user.domain;
 
-import demo.codeexample.enums.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
